Only run the warn when publishing and upgrading
